This commit was manufactured by cvs2svn to create tag 'zabbix'.
git-svn-id: svn://svn.zabbix.com/tags/zabbix@3107 f4e7fd02-d210-4e34-a1e0-d4093fb66cc2

+++ src/zabbix_agent/logfiles.c
@@ -32,12 +32,12 @@
 #include "log.h"
 #include "logfiles.h"
 
-int   process_log(char *filename,int *lastlogsize, char *value)
+int   process_log(char *filename,long *lastlogsize, char *value)
 {
 	FILE	*f;
 	struct stat	buf;
 
-	zabbix_log( LOG_LEVEL_DEBUG, "In process log (%s,%d)", filename, *lastlogsize);
+	zabbix_log( LOG_LEVEL_DEBUG, "In process log (%s,%ld)", filename, *lastlogsize);
 
 	/* Handling of file shrinking */
 	if(stat(filename,&buf) == 0)
@@ -64,7 +64,7 @@ int   process_log(char *filename,int *lastlogsize, char *value)
 
 	if(-1 == fseek(f,*lastlogsize,SEEK_SET))
 	{
-		zabbix_log( LOG_LEVEL_WARNING, "Cannot set postition to [%d] for [%s] [%s]", *lastlogsize, filename, strerror(errno));
+		zabbix_log( LOG_LEVEL_WARNING, "Cannot set postition to [%ld] for [%s] [%s]", *lastlogsize, filename, strerror(errno));
 		snprintf(value,MAX_STRING_LEN-1,"%s","ZBX_NOTSUPPORTED\n");
 		fclose(f);
 		return 1;
